+++ mp3-extract/src/main/java/su/drei/mp3extr/impl/Mp3ThreadedDecoder.java
@@ -96,7 +96,7 @@ public class Mp3ThreadedDecoder extends Mp3Decoder {
         public Boolean call() throws Exception {
             final int framesCnt = bufferSize / (channelsCount * 2);
             // loop over channels in audio stream
-            double[] channelFrames = new double[framesCnt];
+            float[] channelFrames = new float[framesCnt];
             // loop over frames for one channel only
             for (int pos = 0; pos < framesCnt; pos++) {
                 // get two bytes and glue 'em together
