+++ mp3-extract/src/main/java/su/drei/mp3extr/impl/model/ChannelData.java
@@ -2,5 +2,5 @@ package su.drei.mp3extr.impl.model;
 
 public class ChannelData {
     public byte channelNo;
-    public double[] data;
+    public float[] data;
 }
