+++ mp3-extract/src/main/java/su/drei/mp3extr/impl/streaming/PcmChannelMapper.java
@@ -20,7 +20,7 @@ public class PcmChannelMapper implements Function<byte[],Stream<ChannelData>>{
         final int framesCnt = data.length / (channelsCount * 2);
         // loop over channels in audio stream
         for (int chNo = 0; chNo < channelsCount; chNo++) {
-            double[] channelFrames = new double[framesCnt];
+            float[] channelFrames = new float[framesCnt];
             // loop over frames for one channel only
             for (int pos = 0; pos < framesCnt; pos++) {
                 // get two bytes and glue 'em together
