+++ mp3-extract/src/main/java/su/drei/mp3extr/impl/window/BlackmanHarris.java
@@ -2,12 +2,12 @@ package su.drei.mp3extr.impl.window;
 
 public class BlackmanHarris implements IWindowFunc {
 
-    public double[] window(double[] data) {
-        double[] res = new double[data.length];
+    public float[] window(float[] data) {
+        float[] res = new float[data.length];
 
         // Blackman - Harris
         for (int i = 0; i < data.length; i++) {
-            res[i] = data[i] * (0.35875 - 0.48829 * Math.cos(2 * Math.PI * i / (data.length - 1)) + 0.14128 * Math.cos(4 * Math.PI * i / (data.length - 1)) - 0.01168 * Math.cos(6 * Math.PI * i / (data.length - 1)));
+            res[i] = (float) (data[i] * (0.35875 - 0.48829 * Math.cos(2 * Math.PI * i / (data.length - 1)) + 0.14128 * Math.cos(4 * Math.PI * i / (data.length - 1)) - 0.01168 * Math.cos(6 * Math.PI * i / (data.length - 1))));
         }
         return res;
     }
