+++ mp3-extract/src/main/java/su/drei/mp3extr/impl/window/IWindowFunc.java
@@ -2,6 +2,6 @@ package su.drei.mp3extr.impl.window;
 
 public interface IWindowFunc {
 
-    double[] window(double[] input);
+    float[] window(float[] input);
 
 }
